ui: Change ListModel to operate in size_t instead of int.
R=xiyuan@chromium.org
TBR=sky@chromium.org

Review URL: https://chromiumcodereview.appspot.com/10408009

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@137862 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: c6ff734cef5b4f50bc63cf8cefd180d65405dba9

+++ chrome/browser/ui/views/ash/app_list/app_list_model_builder.cc
@@ -151,7 +151,7 @@ void AppListModelBuilder::InsertItemByTitle(app_list::AppListItemModel* item) {
 
   l10n_util::StringComparator<string16> c(collator.get());
   ModelItemSortData data(item);
-  for (int i = special_items_count_; i < model_->item_count(); ++i) {
+  for (size_t i = special_items_count_; i < model_->item_count(); ++i) {
     ModelItemSortData current(model_->GetItemAt(i));
     if (!c(current.key, data.key)) {
       model_->AddItemAt(i, item);
@@ -204,7 +204,7 @@ void AppListModelBuilder::CreateSpecialItems() {
 
 int AppListModelBuilder::FindApp(const std::string& app_id) {
   DCHECK(model_);
-  for (int i = special_items_count_; i < model_->item_count(); ++i) {
+  for (size_t i = special_items_count_; i < model_->item_count(); ++i) {
     ChromeAppListItem* item =
         static_cast<ChromeAppListItem*>(model_->GetItemAt(i));
     if (item->type() != ChromeAppListItem::TYPE_APP)
